Boot lock support
Initial, pretty hacky, support locking down the boot area of the flash
on flashes that support that kind of thing.

In order to not compete with the kernel's SPI driver this is done in
kernel space using a kernel module. In order to not open up too many can
of worms, the kernel modules needs to be signed. Given that userspace
has access to /dev/mem I doubt this will add any security for any
determined person, but its easy enough to add so why not.

Signed-off-by: Christian Svensson <bluecmd@google.com>