Linux as a bootloader
This commit removes the usage of u-boot by forking u-boot's AST2400
startup code and using it to jump straight to a minimal Linux kernel.
This Linux then runs the loader which verifies that the ubifs is healthy
and that the signatures match.

This approach makes more sense for u-bmc in that it allows very early
initialization to be modified to fit whatever the goals are.

This increases the bootloader size with something like 4x but offers
vastly simpler code base and also no longer any outdated software.

Signed-off-by: Christian Svensson <bluecmd@google.com>