Fix boot/loader implementation of kexec
The loader tool can now be used in both the LinuxBoot
and runtime initrd and does not rely on having a kexec
binary in the image as we already got proper kexec_load
support on both arm and arm64. On arm64 the loader tries
kexec_file_load first and falls back to kexec_load if it
did not succeed.

Signed-off-by: Marvin Drees <marvin.drees@9elements.com>