[WIP] Add in-process sshd to replace external sshd
In order to have more control over the sshd implementation
we're moving away from exec.cmd to a proper in-process
version that we have more control over and can test better.
This is mostly still work-in-progress until I figured
out how to properly attach a pty with our in-process
shell.

Signed-off-by: Marvin Drees <marvin.drees@9elements.com>