Use uber-go/zap as our new logger (#227)
* Add new logger

Add better logging by using github.com/uber-go/zap
The logger is implemented as a singleton and only accessed
via a pointer to the struct. It's much easier to read
and handle its output than using the standard Go log
package. Minor linting cleanup done as well.

Signed-off-by: Marvin Drees <marvin.drees@9elements.com>

* Make logger init threadsafe

By using the sync.Once function we guarantee an
atomic init of the logger making it threadsafe to
initialise.

Signed-off-by: Marvin Drees <marvin.drees@9elements.com>

* Remove AddCaller() from logger

This info adds unnecessary overhead to the logging
and was therefore removed

Signed-off-by: Marvin Drees <marvin.drees@9elements.com>

* Remove Sync() from loggers

We don't yet have much use for calling Sync()
on the logger. Might add it again at a later stage.

Signed-off-by: Marvin Drees <marvin.drees@9elements.com>

* Fix typo and add error action

Next to logging the occured error we now
also skip into the next loop when fixing the
IPv6 interfaces.

Signed-off-by: Marvin Drees <marvin.drees@9elements.com>